re-commit from previous and take out installed_packages.txt script garbage
